Add cost_object and cost_object_owner labels to garden_shoot_info
The cost object can be configured in the Billing panel of the Administration tab
of the Project in the Gardener dashboard.
It is stored as an annotation on the project resource.

When determining the cost object of a shoot resource, we do not consider the
project that "contains" the shoot, but rather the project that contains the
secret that is referenced by the secret binding of the shoot.

The additional indirection of the secret binding resource allows to use secrets
from another namespace to create shoot clusters without actually exposing the
secrets to the project owners.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>