Replace golint with golangci-lint and update pipeline golang image (#97)
* Replace golint with golangci-lint, update go version, & upgrade base image

`golint` has been frozen and deprecated. Read more here:
https://github.com/golang/go/issues/38968

This commit replaces it with `golangci-lint`, which makes use of the golang
analysis framework.

Additionally, the image used by the pipeline is upgraded to go 1.22.
The go image used to build the binary and the base image are upgraded to
1.22 and debian12, respectively.

* Eliminate linter findings

Switching to golangci-lint uncovered some issues in the code. This commit
intends to introduce as few code changes as necessary to allow linting to
succeed.