Adapt metric collection to consider CredentialsBinding
CredentialsBindings are meant to replace SecretBindings, which will be
eventually deprecated. However, for the time being, both are valid ways
to set shoot credentials.

This commit updates the metric collection logic to support both
CredentialsBinding and SecretBinding references in shoot specifications.
The cost object for a shoot is now determined from the project of either
reference. This ensures compatibility with either approach.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>