Add support for `CredentialsBinding` resources to identify shoot cost objects (#129)
* Initialise clients and informers for CredentialsBinding

CredentialsBinding belongs to a different garden API group than
SecretBinding. As a result, a new client and informer factory is
necessary.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Add informer for CredentialsBinding to metric collector

Introduce a new informer for CredentialsBindings, required to determine
the correct cost object for a shoot when credentials are set via a
CredentialsBinding reference.

Previously, only SecretBinding references were supported, which only
handled secrets. CredentialsBinding supports both secrets and workload
identities.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Adapt metric collection to consider CredentialsBinding

CredentialsBindings are meant to replace SecretBindings, which will be
eventually deprecated. However, for the time being, both are valid ways
to set shoot credentials.

This commit updates the metric collection logic to support both
CredentialsBinding and SecretBinding references in shoot specifications.
The cost object for a shoot is now determined from the project of either
reference. This ensures compatibility with either approach.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* Clean up some code typos

- Remove unnecessary stopCh argument from setupInformerFactories, as it
  is not used within the function.
- Fix parentheses in gardenManagedSeedInformerFactory.Start call.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>